moved GetFolder method to FilesHelper class
