+++ Objectivity.Test.Automation.Common/DriverContext.cs
@@ -111,7 +111,7 @@ namespace Objectivity.Test.Automation.Common
         {
             get
             {
-                return GetFolder(BaseConfiguration.ScreenShotFolder, this.CurrentDirectory);
+                return FilesHelper.GetFolder(BaseConfiguration.ScreenShotFolder, this.CurrentDirectory);
             }
         }
 
@@ -122,7 +122,7 @@ namespace Objectivity.Test.Automation.Common
         {
             get
             {
-                return GetFolder(BaseConfiguration.DownloadFolder, this.CurrentDirectory);
+                return FilesHelper.GetFolder(BaseConfiguration.DownloadFolder, this.CurrentDirectory);
             }
         }
 
@@ -133,7 +133,7 @@ namespace Objectivity.Test.Automation.Common
         {
             get
             {
-                return GetFolder(BaseConfiguration.PageSourceFolder, this.CurrentDirectory);
+                return FilesHelper.GetFolder(BaseConfiguration.PageSourceFolder, this.CurrentDirectory);
             }
         }
 
@@ -430,41 +430,6 @@ namespace Objectivity.Test.Automation.Common
             {
                 this.SaveScreenshot(new ErrorDetail(this.TakeScreenshot(), DateTime.Now, null), this.ScreenShotFolder, this.TestTitle);
             }
-        }
-
-        /// <summary>
-        /// Gets the folder from app.config as value of given key.
-        /// </summary>
-        /// <param name="appConfigValue">The application configuration value.</param>
-        /// <param name="currentFolder">Directory where assembly files are located</param>
-        /// <returns></returns>
-        public static string GetFolder(string appConfigValue, string currentFolder)
-        {
-            string folder;
-
-            if (string.IsNullOrEmpty(appConfigValue))
-            {
-                folder = currentFolder;
-            }
-            else
-            {
-                if (BaseConfiguration.UseCurrentDirectory)
-                {
-                    folder = currentFolder + appConfigValue;
-                }
-                else
-                {
-                    folder = appConfigValue;
-                }
-
-                if (!Directory.Exists(folder))
-                {
-                    Directory.CreateDirectory(folder);
-                }
-            }
-
-            Logger.Trace(CultureInfo.CurrentCulture, "Folder '{0}'", folder);
-            return folder;
         }
     }
 }
